+++ src/components/pages/NotFound/index.tsx
@@ -1,0 +1,21 @@
+import React, { memo } from 'react';
+import ChangeScreenButton from '../../molecules/ChangeScreenButton';
+import Flex from '../../atoms/Flex';
+import Sprite from '../../atoms/Sprite';
+
+function NotFound(props: {}): JSX.Element {
+  return (
+    <Flex jCenter alCenter column style={{ marginTop: '30%' }}>
+      <Flex txCenter large>
+        This section is currently under construction.
+      </Flex>
+      <Sprite src="https://thumbs.gfycat.com/ParallelQuerulousCottontail.webp" />
+      <Flex txCenter large>
+        We apologize for the inconvenience.
+      </Flex>
+      <ChangeScreenButton to="/">Return to the Pal Park</ChangeScreenButton>
+    </Flex>
+  );
+}
+
+export default memo(NotFound);
