+++ src/store/actionCreator.ts
@@ -1,5 +1,6 @@
 import { CATCH_POKEMON } from './actions';
+import { Pokemon, Action } from '..';
 
-export function catchPokemon(pokemonObj) {
+export function catchPokemon(pokemonObj: Pokemon): Action {
   return { type: CATCH_POKEMON, payload: pokemonObj };
 }
