--- src/store/actions.js
@@ -1,1 +1,0 @@
-export const CATCH_POKEMON = 'CATCH_POKEMON';
