+++ src/store/actions.ts
@@ -1,0 +1,1 @@
+export const CATCH_POKEMON: string = 'CATCH_POKEMON';
